Update app_de.arb
Hello, I'm a native German. I changed everything that sounded weird from a native perspective to sound more appealing and logical. Mostly I edited the sentence structures and every "Sie" to "Du", as this is more common and I saw that the french translation also uses "Tu" which corresponds to "Du". Hope this helps to improve the translation quality!